Add class template
